requirements.txt
google-genai==1.68.0
PyYAML==6.0.3
aiofiles==25.1.0
python-dotenv==1.2.2
